고도 슬라이더 연속적으로 값이 변하게 변경

+++ Assets/02.Scripts/GameController.cs
@@ -10,6 +10,6 @@ public class GameController : MonoBehaviour
     private void Update()
     {
         time += Time.deltaTime;
-        heightBar.SetHeight((int)time);
+        heightBar.SetHeight(time);
     }
 }
